Implemented the Problem module
Problem::Fetcher and Problem::Checker implement the logic for displaying the new
problem that's being solved on the screen and check if the solution is valid.

At the moment, the Problem::Checker isn't working properly since it's not able
to capture the output generated by the test programm written by chat.